// src/_libs/enums.ts
export enum Tabs {
  Home = 'home',
  About = 'about',
  Services = 'services',
  Portfolio = 'portfolio',
  Client = 'client',
  Contact = 'contact',
}

export enum AdminPages {
  Home = 'home',
  Service = 'service',
  Skill = 'skill',
  Design = 'design',
  Video = 'video',
  Review = 'review',
  User = 'user',
  Login = 'login',
  ResetPassword = 'reset',
}
export enum Role {
  USER,
  ADMIN,
}